requirements.txt
pytest
selenium
pytest-xdist
allure-pytest